# Improve 'rate() by ()' queries performance